add camera to WebmakerActivty so it can be used in other activities
