+++ app/src/main/java/org/mozilla/webmaker/WebmakerActivity.java
@@ -1,7 +1,11 @@
 package org.mozilla.webmaker;
 
 import android.app.ActionBar;
+import android.content.Intent;
+import android.graphics.Bitmap;
 import android.os.Bundle;
+import android.os.Environment;
+import android.util.Log;
 import android.view.MenuItem;
 import android.view.ViewConfiguration;
 import android.widget.RelativeLayout;
@@ -9,17 +13,25 @@ import android.widget.RelativeLayout;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import org.mozilla.webmaker.util.Image;
 import org.mozilla.webmaker.view.WebmakerWebView;
 
+import java.io.File;
 import java.lang.reflect.Field;
+import java.util.UUID;
 
 public class WebmakerActivity extends BaseActivity {
 
     public WebmakerWebView view;
+    public String mUuid = UUID.randomUUID().toString();
+    public File mFile = new File(Environment.getExternalStorageDirectory() + File.separator + mUuid + ".jpg");
     protected JSONObject routeParams;
     protected String pageName;
     protected int id, layoutResID, menuResId;
     protected boolean mIsRestart = false;
+    protected final int CAMERA_REQUEST_CODE = 1888;
+    protected final int MEDIA_REQUEST_CODE = 1889;
+
 
     public WebmakerActivity(String pageName, int id, int layoutResID, int menuResId) {
         this.pageName = pageName;
@@ -122,5 +134,52 @@ public class WebmakerActivity extends BaseActivity {
                 return true;
         }
         return super.onOptionsItemSelected(item);
+    }
+
+    /**
+     * Dispatches a new image capture intent.
+     */
+    public void dispatchCameraIntent() {
+        Intent intent = Image.getCameraIntent(mFile);
+        startActivityForResult(intent, CAMERA_REQUEST_CODE);
+    }
+
+    /**
+     * Dispatches a new media gallery intent.
+     */
+    public void dispatchMediaIntent() {
+        Intent intent = Image.getMediaStoreIntent(mFile);
+        startActivityForResult(intent, MEDIA_REQUEST_CODE);
+    }
+
+    /**
+     * Result handler for the image capture activity.
+     */
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        Bitmap bitmap = null;
+
+        // Return if user exited without capturing / selecting an image
+        if (resultCode == 0) return;
+
+        // Handle camera activity
+        if (requestCode == CAMERA_REQUEST_CODE) {
+            bitmap = Image.decodeBitmapFromFile(mFile.getAbsolutePath(), 600, 600);
+        }
+
+        // Handle media (gallery) activity
+        if (requestCode == MEDIA_REQUEST_CODE) {
+            bitmap = Image.decodeBitmapFromMediaStore(data.getData(), 600, this);
+        }
+
+        // @todo Handle error
+        if (bitmap == null) {
+            Log.e("DATAURI:ERROR", "Bitmap is null");
+            return;
+        }
+
+        // Convert bitmap to data uri and forward to JS
+        String uri = Image.createDataUriFromBitmap(bitmap, 80);
+        view.load("javascript: window.imageReady && window.imageReady('" + uri + "')", null);
     }
 }
